Change Octaves to be u8 instead of usize
u8 allows up to 256 octaves, but most of the time, the number of octaves will probably be less than 10. Don't need full pointer width to store the number of octaves.

+++ src/noise_fns/generators/fractals.rs
@@ -11,7 +11,7 @@ use crate::Seedable;
 
 /// Trait for `MultiFractal` functions
 pub trait MultiFractal {
-    fn set_octaves(self, octaves: usize) -> Self;
+    fn set_octaves(self, octaves: u8) -> Self;
 
     fn set_frequency(self, frequency: f64) -> Self;
 
@@ -20,11 +20,11 @@ pub trait MultiFractal {
     fn set_persistence(self, persistence: f64) -> Self;
 }
 
-fn build_sources<Source>(seed: u32, octaves: usize) -> Vec<Source>
+fn build_sources<Source>(seed: u32, octaves: u8) -> Vec<Source>
 where
     Source: Default + Seedable,
 {
-    let mut sources = Vec::with_capacity(octaves);
+    let mut sources = Vec::with_capacity(octaves as usize);
     for x in 0..octaves {
         let source = Source::default();
         sources.push(source.set_seed(seed + x as u32));
